Put the enum into the new class and fix some closure issues.
