+++ src/xcode/ENA/ENA/Source/Scenes/DynamicTableViewController/Views/DynamicTableViewConsentCell.swift
@@ -169,15 +169,6 @@ class DynamicTableViewConsentCell: UITableViewCell {
 		style.lineSpacing = 10
 		flagString.addAttribute(.paragraphStyle, value: style, range: NSRange(location: 0, length: flagString.length))
 				
-		self.flagIconsLabel.attributedText = flagString
-	}
-}
-
-extension NSTextAttachment {
-	func setImageHeight(height: CGFloat) {
-		guard let image = image else { return }
-		let ratio = image.size.width / image.size.height
-
-		bounds = CGRect(x: bounds.origin.x, y: bounds.origin.y, width: ratio * height, height: height)
+		flagIconsLabel.attributedText = flagString
 	}
 }
